+++ src/clients/GitHub.php
@@ -44,7 +44,6 @@ class GitHub extends OAuth2
      * {@inheritdoc}
      */
     public $accessTokenLocation = OAuth2::ACCESS_TOKEN_LOCATION_HEADER;
-
     /**
      * {@inheritdoc}
      */