+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 authclient extension Change Log
 2.2.16 under development
 ------------------------
 
-- no changes in this release.
+- Enh #387: Use appropriate exception if client does not exist (eluhr)
 
 
 2.2.15 December 16, 2023